Adds basic http authentication. Adds fields validations.

+++ src/Controller/NoteController.php
@@ -25,6 +25,7 @@ class NoteController
     /**
      * @Route("/notes", name="add_note", methods={"POST"})
      * @param Request $request
+     *
      * @return JsonResponse
      */
     public function add(Request $request): JsonResponse
@@ -33,12 +34,15 @@ class NoteController
 
         $title = $data['title'];
         $note = $data['note'];
-        $user = $data['user'];
 
-        //@TODO: decide if search of user by id or email
-        $user = $this->userRepository->findOneBy(['id' => $user]);
+        $user = $this->isUserAuthorizedByBasicHttp($request);
+        if (empty($user)) {
+            return new JsonResponse(['status' => 'You are not authorized!'], Response::HTTP_UNAUTHORIZED);
+        }
 
-        if (empty($title) || empty($note) || empty($user)) {
+
+        if (empty($title) || empty($note)) {
+            //@TODO: return bad request
             throw new NotFoundHttpException('Expecting mandatory parameters!');
         }
 
@@ -49,12 +53,18 @@ class NoteController
 
     /**
      * @Route("/notes", name="get_all_user_notes", methods={"GET"})
+     * @param Request $request
+     *
+     * @return JsonResponse
      */
-    public function getAllByUserId(): JsonResponse
+    public function getAllByUserId(Request $request): JsonResponse
     {
-        //@TODO: receive user id as param. match notes for a specific user
-        $notes = $this->noteRepository->findAllByUserId();
-//        $notes = $this->noteRepository->findAllByUserId(['id' => $id]);
+        $user = $this->isUserAuthorizedByBasicHttp($request);
+        if (empty($user)) {
+            return new JsonResponse(['status' => 'You are not authorized!'], Response::HTTP_UNAUTHORIZED);
+        }
+
+        $notes = $this->noteRepository->findAllByUserId($user->getId());
 
         $data = [];
         foreach ($notes as $note) {
@@ -87,9 +97,17 @@ class NoteController
     public function update(int $id, Request $request): JsonResponse
     {
         $note = $this->noteRepository->findOneBy(['id' => $id]);
-        //@TODO: validate the note exists.
-        //@TODO: Validate the user is the owner of the note
 
+        if (empty($note)) {
+            return new JsonResponse(['status' => 'The note does not exist'], Response::HTTP_NOT_FOUND);
+        }
+
+        $user = $this->isUserAuthorizedByBasicHttp($request);
+        if (empty($user) || $note->getUser()->getEmail() != $user->getEmail()) {
+            return new JsonResponse(['status' => 'You are not authorized!'], Response::HTTP_UNAUTHORIZED);
+        }
+
+        //@TODO: can a note change the ownership to another user? Assume no for now
         $data = json_decode($request->getContent(), true);
 
         empty($data['title']) ? true : $note->setTitle($data['title']);
@@ -105,18 +123,46 @@ class NoteController
     /**
      * @Route("/notes/{id}", name="delete_note", methods={"DELETE"})
      * @param int $id
+     * @param Request $request
      *
      * @return JsonResponse
      */
-    public function delete(int $id): JsonResponse
+    public function delete(int $id, Request $request): JsonResponse
     {
         $note = $this->noteRepository->findOneBy(['id' => $id]);
+        if (empty($note)) {
+            return new JsonResponse(['status' => 'The note does not exist'], Response::HTTP_NOT_FOUND);
+        }
 
-        //@TODO: validate the note exists
-        //@TODO: Validate the user is the owner of the note
+        $user = $this->isUserAuthorizedByBasicHttp($request);
+        if (empty($user) || $note->getUser()->getEmail() != $user->getEmail()) {
+            return new JsonResponse(['status' => 'You are not authorized!'], Response::HTTP_UNAUTHORIZED);
+        }
 
         $this->noteRepository->removeNote($note);
 
-        return new JsonResponse(['status' => 'Note deleted'], Response::HTTP_OK);
+        return new JsonResponse(['status' => 'Note deleted'], Response::HTTP_NO_CONTENT);
+    }
+
+
+    /**
+     * Returns the user if the access is granted. Null otherwise
+     * @param Request $request
+     *
+     * @return \App\Entity\User|null
+     */
+    private function isUserAuthorizedByBasicHttp(Request $request) {
+        //@TODO: different types of authorization? Explode, and with a factory get the correct manager based on the type (basic, bearer, etc...)
+        $auth = $request->headers->get("Authorization");
+        if (empty($auth)) return null;
+
+        $auth = explode(" ", $auth);
+        $auth = base64_decode($auth[1]);
+        $auth = explode(":", $auth);
+
+        $user = $this->userRepository->findOneBy(['email' => $auth[0]]);
+
+        //@TODO:password should be encoded in DB.
+        return !empty($user) && $user->getPassword() == $auth[1] ? $user : null;
     }
 }
